Directional Aiming
The location of the AttackPoint (which is the Vector2 where the attack circle/box will spawn from) is now dependent on the character's input, both horizontal and vertical. If you aim vertically (up or down), or both vertcially and horizontally, once the character's input returns to 0, the attackpoint will return to the last horizontal input that was recorded.